mqtt/requirements.txt
absl-py==0.1.13
astor==0.6.2
astroid==1.6.3
backports.functools-lru-cache==1.5
backports.weakref==1.0.post1
bleach==1.5.0
configparser==3.5.0
cycler==0.10.0
enum34==1.1.6
funcsigs==1.0.2
futures==3.2.0
gast==0.2.0
grpcio==1.10.0
h5py==2.7.1
html5lib==0.9999999
isort==4.3.4
Keras==2.1.5
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
Markdown==2.6.11
matplotlib==2.2.2
mccabe==0.6.1
mock==2.0.0
more-itertools==4.1.0
numpy==1.14.2
paho-mqtt==1.4.0
pandas==0.23.0
pbr==4.0.1
Pillow==5.1.0
protobuf==3.5.2.post1
pylint==1.8.4
pyparsing==2.2.0
python-dateutil==2.7.2
pytz==2018.3
PyYAML==3.12
scikit-learn==0.19.1
scipy==1.0.1
singledispatch==3.4.0.3
six==1.11.0
subprocess32==3.2.7
tensorboard==1.7.0
tensorflow==1.7.0
termcolor==1.1.0
Werkzeug==0.14.1
wrapt==1.10.11
